backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.50.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.2
httpx>=0.27.0,<0.28.0
